requirements.txt
btchip-python==0.1.32
base58==2.1.1
bip32==3.3

trezor[hidapi]==0.13.4
